feat(rag): introduce retrieval module with registry and retrieval_method param
Creates app/rag/retrieval/ as the home for all retrieval implementations.
Adding a new retrieval method only requires adding a file here and one entry
in the registry.

- Move agent.py -> retrieval/agent.py; add get_context_via_agent to it
  (the full agentic orchestration now lives alongside the agent it uses)
- retrieval/__init__.py: RETRIEVAL_REGISTRY (str -> Callable) + public
  retrieve_context_from_file(user_id, file_id, query, llm, retrieval_method)
  which dispatches to the registered function, independent of implementation
- context_service.py: now only shared utilities (_lookup_drive_file,
  format_chunks_as_context, _validate_files_ready)
- retrieve_context() in app/lib/rag gains retrieval_method="agentic" and
  passes it through to retrieve_context_from_file
